all rake commands work

+++ db/migrate/20221101190305_create_transactions.rb
@@ -1,7 +1,7 @@
 class CreateTransactions < ActiveRecord::Migration[5.2]
   def change
     create_table :transactions do |t|
-      t.integer :credit_card_number
+      t.bigint :credit_card_number
       t.string :cc_expiration
       t.integer :result
       t.references :invoice, foreign_key: true
